separated lists, created main navigation

+++ src/Controller/ExpenseAdminController.php
@@ -4,7 +4,6 @@ namespace App\Controller;
 
 use App\Entity\Expense;
 use App\Form\ExpenseFormType;
-use App\Repository\CarRepository;
 use App\Repository\ExpenseRepository;
 use Doctrine\ORM\EntityManagerInterface;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
@@ -14,17 +13,34 @@ use Symfony\Component\Routing\Annotation\Route;
 
 class ExpenseAdminController extends AbstractController
 {
-    #[Route('/admin/expense/new/{car}', name: 'app_expense_new')]
-    public function new(EntityManagerInterface $em, CarRepository $carRepo, Request $request, $car): Response
+    #[Route('/admin/expense/list', name: 'app_expense_list')]
+    public function list(ExpenseRepository $expRepo)
     {
-        $carObj = $carRepo->find($car);
-        if (!$carObj->hasUser($this->getUser())) {
-            $this->redirectToRoute('app_car_list');
-        }
+        /** @var User $user */
+        $user = $this->getUser();
+        $car = $user->getCar();
+
+        $expenses = $expRepo->findByCar($car);
+
+        return $this->render(
+            'admin/expense/list.html.twig',
+            [
+                'car' => $car,
+                'expenses' => $expenses,
+            ]
+        );
+    }
+
+    #[Route('/admin/expense/new', name: 'app_expense_new')]
+    public function new(EntityManagerInterface $em, Request $request): Response
+    {
+        /** @var User $user */
+        $user = $this->getUser();
+        $car = $user->getCar();
 
         $expense = new Expense();
         $expense->setUser($this->getUser());
-        $expense->setCar($carObj);
+        $expense->setCar($car);
 
         $form = $this->createForm(ExpenseFormType::class, $expense);
 
@@ -36,14 +52,14 @@ class ExpenseAdminController extends AbstractController
 
             $this->addFlash('success', 'Expense created!');
 
-            return $this->redirectToRoute('app_car_show', ['car' => $carObj->getId()]);
+            return $this->redirectToRoute('app_car_show');
         }
 
         return $this->render(
             'admin/expense/new.html.twig',
             [
                 'expenseForm' => $form->createView(),
-                'car' => $carObj,
+                'car' => $car,
             ]
         );
     }
@@ -51,7 +67,7 @@ class ExpenseAdminController extends AbstractController
     #[Route('/admin/expense/edit/{expense}', name: 'app_expense_edit')]
     public function edit(EntityManagerInterface $em, Request $request, Expense $expense): Response
     {
-        $carObj = $expense->getCar();
+        $car = $expense->getCar();
         $form = $this->createForm(ExpenseFormType::class, $expense);
 
         if ($this->getUser() !== $expense->getUser()) {
@@ -66,31 +82,37 @@ class ExpenseAdminController extends AbstractController
 
                 $this->addFlash('success', 'Expense updated!');
 
-                return $this->redirectToRoute('app_car_show', ['car' => $carObj->getId()]);
+                return $this->redirectToRoute('app_car_show');
             }
 
             return $this->render(
                 'admin/expense/edit.html.twig',
                 [
                     'expenseForm' => $form->createView(),
-                    'car' => $carObj,
+                    'car' => $car,
                 ]
             );
         }
 
-        return $this->redirectToRoute('app_trip_list', ['car' => $carObj->getId()]);
+        return $this->redirectToRoute('app_trip_list', ['car' => $car->getId()]);
     }
 
     #[Route('/admin/expense/delete/{expense}', name: 'app_expense_delete')]
     public function delete(EntityManagerInterface $em, ExpenseRepository $expenseRepo, $expense)
     {
         $expense = $expenseRepo->find($expense);
-        $car = $expense->getCar();
+
+        if ($this->getUser() !== $expense->getUser()) {
+            $this->addFlash('error', 'You can only delete your own expenses.');
+
+            return $this->redirectToRoute('app_trip_list');
+        }
+
         $em->remove($expense);
         $em->flush();
 
         $this->addFlash('success', 'Expense deleted.');
 
-        return $this->redirectToRoute('app_car_show', ['car' => $car->getId()]);
+        return $this->redirectToRoute('app_car_show');
     }
 }
